debug follow friend

+++ src/main/java/com/webcloud/tinygram/MyApi.java
@@ -105,9 +105,10 @@ public class MyApi {
         if (list == null) {
             list = new ArrayList<String>();
         }
+
         list.add(follow);
         e.setProperty("following", list);
-        e.setProperty("cptFollowing", (int) e.getProperty("cptFollowing")+1);
+        e.setProperty("cptFollowing", (long) e.getProperty("cptFollowing")+1);
 
 
         Key key2 = KeyFactory.createKey("Friend", follow);
@@ -118,7 +119,7 @@ public class MyApi {
         }
         listFollowers.add(email);
         e2.setProperty("followers", listFollowers);
-        e2.setProperty("cptFollower", (int) e.getProperty("cptFollower")+1);
+        e2.setProperty("cptFollower", (long) e.getProperty("cptFollower")+1);
 
         datastore.put(e);
         datastore.put(e2);
